Package name fixup
I am not a master of java packages.
The build was failing because WebServer was ambiguous (would have been
nice if the compiler had said that, instead of "cannot find symbol.")
Adding a package name to my project clarified that.
IntelliJ is still complaining about my package names, nomatter if I try
adding what I think the "right" name is. I'm going to play with
resolving that some....