Delayed messaging
Got delayed messaging working; I'm going to be using this as the
scheduler for polling the registered websites. Working on a
RabbitPublisher class since both the webserver and taskconsumer will
need to publish messages.
I want t keep the rabbit objects contained within this class, but I need
to register a handler on the task consumer... I can dig up the rabbit
interface object, and take an instance of that as a parameter. It'll be
good for testing too.