Task Consumer Testing
Finished the task consumer test, which posts one message, consumes it,
detects a website change and sends an email. I need to mock the emailer
portion, as right now I'm actually checking email to verify it.

Found an error where I wasn't decoding the byte-stream that RabbitMQ
holds as UTF-8.