Unit test fix
I had the server sending the response before posting the task, since
that can happen later and there's no need to block the client response.
However, that makes the thread.interupt() in the unit test block that
asynchronous work.
Checking in now because I'm about to integrate a logger. I'll also be
fixing the error handling of the main task along with it.