Log4j integration, exception refactoring
Integrated log4j on the server. Put in a few log messages. Set up a
default configuration.
Refactored the exception handling in the server to be a bit more
realistic.
Next I'm going to extract the interfaces for the dependencies so that I
can inject them so I can properly test the server.